commitlint(footer-refs-validity): fix rule
Previously we passed everything to the function using body
but the problem is that for some reason when we use a link
that has a hashtag before a number for reference, commitlint
puts that line in the footer instead of body. So by passing
both footer and body to the function and concatenating them
the problem is fixed.